+++ src/SipConnector.ts
@@ -265,6 +265,7 @@ type TSet = ({
   password?: string;
 }) => Promise<boolean>;
 
+type TDegradationPreference = 'maintain-framerate' | 'maintain-resolution' | 'balanced';
 type TCall = ({
   number,
   mediaStream,
@@ -279,6 +280,7 @@ type TCall = ({
   iceServers?: RTCIceServer[];
   videoMode?: 'sendrecv' | 'sendonly' | 'recvonly';
   audioMode?: 'sendrecv' | 'sendonly' | 'recvonly';
+  degradationPreference?: TDegradationPreference;
 }) => Promise<RTCPeerConnection>;
 
 type TDisconnect = () => Promise<void>;
@@ -290,6 +292,7 @@ type TParametersAnswerToIncomingCall = {
   iceServers?: RTCIceServer[];
   videoMode?: 'sendrecv' | 'sendonly' | 'recvonly';
   audioMode?: 'sendrecv' | 'sendonly' | 'recvonly';
+  degradationPreference?: TDegradationPreference;
 };
 
 type TAnswerToIncomingCall = (
@@ -491,6 +494,7 @@ export default class SipConnector {
       deleteExisting: boolean;
       addMissing: boolean;
       forceRenegotiation: boolean;
+      degradationPreference?: TDegradationPreference;
     }
   ): Promise<void> {
     if (!this.session) {
@@ -559,10 +563,12 @@ export default class SipConnector {
       isNeedReinvite = true,
       isP2P = false,
       maxBitrate,
+      degradationPreference,
     }: {
       isNeedReinvite?: boolean;
       isP2P?: boolean;
       maxBitrate?: number;
+      degradationPreference?: TDegradationPreference;
     } = {}
   ): Promise<void | MediaStream> {
     const session = this.establishedSession;
@@ -588,7 +594,11 @@ export default class SipConnector {
         extraHeaders: preparatoryHeaders,
       })
       .then(() => {
-        return session.startPresentation(streamPresentationCurrent, isNeedReinvite);
+        return session.startPresentation(
+          streamPresentationCurrent,
+          isNeedReinvite,
+          degradationPreference
+        );
       })
       // @ts-ignore
       .then(() => {
@@ -666,9 +676,12 @@ export default class SipConnector {
 
       const callerData = this.remoteCallerData;
 
-      session.on(FAILED, () => {
+      session.on(FAILED, ({ originator }) => {
         this.removeIncomingSession();
-        this._uaEvents.trigger(FAILED_INCOMING_CALL, callerData);
+
+        if (originator !== ORIGINATOR_LOCAL) {
+          this._uaEvents.trigger(FAILED_INCOMING_CALL, callerData);
+        }
       });
 
       this._uaEvents.trigger(INCOMING_CALL, callerData);
@@ -990,6 +1003,7 @@ export default class SipConnector {
     iceServers,
     videoMode,
     audioMode,
+    degradationPreference,
   }) => {
     return new Promise((resolve, reject) => {
       this._connectionConfiguration.number = number;
@@ -1005,6 +1019,7 @@ export default class SipConnector {
         eventHandlers: this._sessionEvents.triggers,
         videoMode,
         audioMode,
+        degradationPreference,
         pcConfig: {
           iceServers,
         },
@@ -1023,6 +1038,7 @@ export default class SipConnector {
     iceServers,
     videoMode,
     audioMode,
+    degradationPreference,
   }): Promise<RTCPeerConnection> => {
     return new Promise((resolve, reject) => {
       if (!this.isAvailableIncomingCall) {
@@ -1065,6 +1081,7 @@ export default class SipConnector {
         extraHeaders,
         videoMode,
         audioMode,
+        degradationPreference,
         mediaStream: preparedMediaStream,
         pcConfig: {
           iceServers,
@@ -1524,7 +1541,7 @@ export default class SipConnector {
     return this.waitSession(CHANNELS);
   }
 
-  sendChannels({ inputChannels, outputChannels }: TChannels) {
+  sendChannels({ inputChannels, outputChannels }: TChannels): Promise<void> {
     if (!this.session) {
       throw new Error('No session established');
     }
@@ -1536,7 +1553,7 @@ export default class SipConnector {
       headerOutputChannels,
     ];
 
-    this.session.sendInfo(CONTENT_TYPE_CHANNELS, undefined, { extraHeaders });
+    return this.session.sendInfo(CONTENT_TYPE_CHANNELS, undefined, { extraHeaders });
   }
 
   sendMediaState(