+++ src/__tests__/incomingCallStatuses.ts
@@ -63,9 +63,9 @@ describe('incoming call statuses', () => {
 
     return promiseCallStatuses.then(() => {
       expect(sipConnector.isCallActive).toBe(true);
-      expect(mockFnConnecting).toBeCalledTimes(1);
-      expect(mockFnAccepted).toBeCalledTimes(1);
-      expect(mockFnConfirmed).toBeCalledTimes(1);
+      expect(mockFnConnecting).toHaveBeenCalledTimes(1);
+      expect(mockFnAccepted).toHaveBeenCalledTimes(1);
+      expect(mockFnConfirmed).toHaveBeenCalledTimes(1);
     });
   });
 });